Fix date parsing in name of downloaded files (workaround...read below)
It looks like the date format in the name of the downloaded csv files
can change which breaks the merging afterwards.
With Selenium it does not seem to be possible to set the name of
the files that we want to download. One way to work around this,
assuming changing date format, is to look for the most recent file
after each download and rename it to match the currently requested
date, while taking care of not renaming older files if downloads
failed (or simply there's no data) inbetweeen different requests.